\name{dataLym}
\Rdversion{1.4}
\alias{dataLym}
\docType{data}
\title{
Additional data for the lymphoma microarray chip (exprLym)
}
\description{
The dataset contains additional data for the exprLym dataset. It includes p-values for the gene
expression, p-values for the survival data, an example score etc.
}
\usage{data(dataLym)}
\examples{
data(dataLym)
str(dataLym)
}
\references{A. Rosenwald et al. (2002). The use of molecular profiling to predict survival after chemotherapy for diffuse large-B-cell lymphoma. \emph{N Engl J Med}, 346(25), 1937-1947.}
\keyword{datasets}