\name{DLBCL-package}
\Rdversion{1.2}
\alias{DLBCL-package}
\alias{DLBCL}
\docType{package}
\title{
Routines for the functional analysis of biological networks
}
\description{
This package provides an expression set on diffuse large B-cell lymphoma and an protein-protein interaction network obtained from HPRD release 6. It accompanies the BioNet packages as example data.
For further information on the data and the BioNet package, see references below.
}
\details{
\tabular{ll}{
Package: \tab DLBCL\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2010-03-26\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Marcus Dittrich, Daniela Beisser

Maintainer: Marcus Dittrich <marcus.dittrich@biozentrum.uni-wuerzburg.de>
}
\references{M. T. Dittrich, G. W. Klau, A. Rosenwald, T. Dandekar and T. Mueller (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. \emph{(ISMB2008) Bioinformatics} 24: 13. i223-i231 Jul.


D. Beisser, G. W. Klau, T. Dandekar, T. Mueller and M. Dittrich (2009) BioNet: an R-package for the Functional Analysis of Biological Networks. \emph{Bioinformatics}.

A. A. Alizadeh, M. B. Eisen, R. E. Davis et al. (2000) Distinct types of diffuse large B-cell lymphoma identified by gene expression profiling. \emph{Nature} 403: 503-11.
}

