\name{exprLym}
\Rdversion{1.4}
\alias{exprLym}
\docType{data}
\title{
Expression set diffuse large B-cell lymphomas
}
\description{
The dataset contains an expression set on diffuse large B-cell lymphoma. It accompanies the BioNet packages as example data.
For further information on the data and the BioNet package see:

M. T. Dittrich, G. W. Klau, A. Rosenwald, T. Dandekar and T. Mueller (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. \emph{(ISMB2008) Bioinformatics} 24: 13. i223-i231 Jul.


D. Beisser, G. W. Klau, T. Dandekar, T. Mueller and M. Dittrich (2009) BioNet: an R-package for the Functional Analysis of Biological Networks. \emph{Bioinformatics}.


A. A. Alizadeh, M. B. Eisen, R. E. Davis et al. (2000) Distinct types of diffuse large B-cell lymphoma identified by gene expression profiling. \emph{Nature} 403: 503-11.
}
\usage{data(exprLym)}
\examples{
data(exprLym)
exprs(exprLym)[1:10,]
}
\keyword{datasets}