\name{interactome}
\Rdversion{1.4}
\alias{interactome}
\docType{data}
\title{
Human protein-protein interaction network
}
\description{
The dataset contains the human proteome, extracted from the Human Protein Reference Database
(HPRD) from 2006 that is used in the ABC GCB diffuse large B-cell lymphma analysis.
 The format of the dataset is a graph object.
}
\usage{data(interactome)}
\references{
M. T. Dittrich, G. W. Klau, A. Rosenwald, T. Dandekar, T. Mueller (2008) Identifying functional modules in protein-protein interaction networks: an integrated exact approach. \emph{(ISMB2008) Bioinformatics}, 24: 13. i223-i231 Jul.
}
\examples{
data(interactome)
interactome
}
\keyword{datasets}